fix some clippy warnings of newer rust version

+++ block_backend/src/qcow2/table.rs
@@ -188,12 +188,12 @@ impl Qcow2Table {
 
     pub fn update_l2_table(&mut self, l2_table_entry: Rc<RefCell<CacheTable>>) -> Result<()> {
         let l2_entry_addr = l2_table_entry.borrow().addr;
-        if self.l2_table_cache.contains_keys(l2_entry_addr as u64) {
+        if self.l2_table_cache.contains_keys(l2_entry_addr) {
             self.l2_table_cache.cache_map.remove(&l2_entry_addr);
         }
         if let Some(replaced_entry) = self
             .l2_table_cache
-            .lru_replace(l2_entry_addr as u64, l2_table_entry)
+            .lru_replace(l2_entry_addr, l2_table_entry)
         {
             let borrowed_entry = replaced_entry.borrow();
             // Flush the dirty entry.
